+++ lingotrace/core/capabilities.py
@@ -12,6 +12,8 @@ PUBLIC_CAPABILITY_IDS = {
     "source_notes",
     "review_materials",
     "review_queue",
+    "review_lifecycle_migration",
+    "vocab_consolidation",
     "speaking_cards",
     "review_rollover",
     "total_training_dashboard",